requirements.txt
contourpy==1.1.1
cycler==0.12.0
fonttools==4.43.0
glcontext==2.5.0
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.8.0
moderngl==5.8.2
numpy==1.26.0
packaging==23.2
pandas==2.1.2
Pillow==10.0.1
plotly==5.18.0
pluggy==1.3.0
pygame==2.5.2
PyGLM==2.7.0
pyparsing==3.1.1
python-dateutil==2.8.2
ruff==0.0.291
six==1.16.0
tenacity==8.2.3
tzdata==2023.3
